feat:ground effect

+++ back/src/recipe/recipe.service.ts
@@ -3,11 +3,12 @@ import { CreateRecipeDto } from './dto/create-recipe.dto';
 import { UpdateRecipeDto } from './dto/update-recipe.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Category } from 'src/entities/category.entity';
-import { Like, Repository } from 'typeorm';
+import { Repository } from 'typeorm';
 import { Market } from 'src/entities/market.entity';
 import { Icon } from 'src/entities/icon.entity';
 import { Creation } from 'src/entities/creation.entity';
 import { Ingredient } from 'src/entities/ingredient.entity';
+import { ConfigService } from '@nestjs/config';
 
 @Injectable()
 export class RecipeService {
@@ -22,57 +23,60 @@ export class RecipeService {
     private creationRepository: Repository<Creation>,
     @InjectRepository(Ingredient)
     private ingredientRepository: Repository<Ingredient>,
+    private configService: ConfigService,
   ) {}
 
-  async create(createRecipeDto: CreateRecipeDto) {
-    console.log(createRecipeDto);
-    const findicategory = await this.categoryRepository.findOne({
-      where: { categoryName: createRecipeDto.category },
-    });
+  async create(createRecipeDto: CreateRecipeDto, session: any) {
+    try {
+      if (session.user !== this.configService.get<string>(`ADMINNAME`)) {
+        throw Error('not admin');
+      }
+      const findicategory = await this.categoryRepository.findOne({
+        where: { categoryName: createRecipeDto.category },
+      });
 
-    let findicon = await this.iconRepository.findOne({
-      where: { itemCode: createRecipeDto.itemCode },
-    });
-    if (!findicon) {
-      findicon = await this.iconRepository.findOne({
-        where: {
+      let findicon = await this.iconRepository.findOne({
+        where: { itemCode: createRecipeDto.itemCode },
+      });
+      if (!findicon) {
+        findicon = await this.iconRepository.findOne({
+          where: {
+            icon: createRecipeDto.icon.slice(
+              'https://cdn-lostark.game.onstove.com/efui_iconatlas/'.length,
+            ),
+          },
+        });
+      }
+      if (!findicon) {
+        findicon = this.iconRepository.create({
           icon: createRecipeDto.icon.slice(
             'https://cdn-lostark.game.onstove.com/efui_iconatlas/'.length,
           ),
-        },
+          itemCode: createRecipeDto.itemCode,
+        });
+        await this.iconRepository.save(findicon);
+      }
+
+      let findmarket = await this.marketRepository.findOne({
+        where: { itemCode: createRecipeDto.itemCode },
       });
-    }
-    if (!findicon) {
-      findicon = this.iconRepository.create({
-        icon: createRecipeDto.icon.slice(
-          'https://cdn-lostark.game.onstove.com/efui_iconatlas/'.length,
-        ),
+      if (!findmarket) {
+        findmarket = await this.marketRepository.findOne({
+          where: { itemCode: 1 },
+        });
+      }
+
+      const creation = this.creationRepository.create({
+        name: createRecipeDto.itemName,
         itemCode: createRecipeDto.itemCode,
+        createBundle: createRecipeDto.createBundle,
+        energy: createRecipeDto.energy,
+        createTime: createRecipeDto.createTime,
+        createCost: createRecipeDto.cost,
+        market: findmarket,
+        icon: findicon,
+        category: findicategory,
       });
-      await this.iconRepository.save(findicon);
-    }
-
-    let findmarket = await this.marketRepository.findOne({
-      where: { itemCode: createRecipeDto.itemCode },
-    });
-    if (!findmarket) {
-      findmarket = await this.marketRepository.findOne({
-        where: { itemCode: 1 },
-      });
-    }
-
-    const creation = this.creationRepository.create({
-      name: createRecipeDto.itemName,
-      itemCode: createRecipeDto.itemCode,
-      createBundle: createRecipeDto.createBundle,
-      energy: createRecipeDto.energy,
-      createTime: createRecipeDto.createTime,
-      createCost: createRecipeDto.cost,
-      market: findmarket,
-      icon: findicon,
-      category: findicategory,
-    });
-    try {
       await this.creationRepository.save(creation);
       // 재료 추가 반복문
       for (const ingredient of createRecipeDto.ingredient) {
@@ -89,19 +93,23 @@ export class RecipeService {
         await this.ingredientRepository.save(ingredientCreate);
       }
       //재료 추가 반복문 끝
+      return { itemId: creation.id };
     } catch (error) {
       console.log(error.message);
       if (error.message.includes('Duplicate')) {
-        return { result: 'duplication Item' };
-      } else {
-        return `${creation.name} 추가중 알수없는 오류가 발생했습니다.`;
+        return { statusCode: 405, result: 'duplication Item' };
+      } else if (error.message.includes('not admin')) {
+        return { statusCode: 401, result: 'not admin' };
       }
+      return { statusCode: 401, result: 'fail' };
     }
-    return { itemId: creation.id };
   }
 
-  async update(id: number, updateRecipeDto: UpdateRecipeDto) {
+  async update(id: number, updateRecipeDto: UpdateRecipeDto, session: any) {
     try {
+      if (session.user !== this.configService.get<string>(`ADMINNAME`)) {
+        throw Error('not admin');
+      }
       const findicategory = await this.categoryRepository.findOne({
         where: { categoryName: updateRecipeDto.category },
       });
@@ -172,21 +180,29 @@ export class RecipeService {
     } catch (error) {
       console.log(error.message);
       if (error.message.includes('Duplicate')) {
-        return { result: 'duplication Item' };
+        return { statusCode: 405, result: 'duplication Item' };
+      } else if (error.message.includes('not admin')) {
+        return { statusCode: 401, result: 'not admin' };
       }
-      return { result: 'fail' };
+      return { statusCode: 400, result: 'fail' };
     }
 
     return { itemId: id };
   }
 
-  async remove(id: number) {
+  async remove(id: number, session: any) {
     try {
+      if (session.user !== this.configService.get<string>(`ADMINNAME`)) {
+        throw Error('not admin');
+      }
       this.ingredientRepository.delete({ creation: { id: id } });
       this.creationRepository.delete({ id: id });
       return { result: 'ok' };
     } catch (error) {
-      return { result: 'fail' };
+      if (error.message.includes('not admin')) {
+        return { statusCode: 401, result: 'not admin' };
+      }
+      return { statusCode: 400, result: 'fail' };
     }
   }
 }
